# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LifeOS Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type Category = typeof Category[keyof typeof Category];


export const Category = {
  career: 'career',
  learning: 'learning',
  health: 'health',
  mind: 'mind',
  routine: 'routine',
  personal: 'personal',
} as const;

export type Frequency = typeof Frequency[keyof typeof Frequency];


export const Frequency = {
  daily: 'daily',
  weekdays: 'weekdays',
  weekly: 'weekly',
} as const;

export type GoalStatus = typeof GoalStatus[keyof typeof GoalStatus];


export const GoalStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  completed: 'completed',
  skipped: 'skipped',
} as const;

export interface Goal {
  id: number;
  name: string;
  category: Category;
  targetValue: number;
  unit: string;
  frequency: Frequency;
  startDate: string;
  /** @nullable */
  endDate: string | null;
  active: boolean;
  /**
     * @minimum 1
     * @maximum 3
     */
  priority: number;
}

export interface GoalInput {
  /** @minLength 1 */
  name: string;
  category: Category;
  /** @exclusiveMinimum 0 */
  targetValue: number;
  /** @minLength 1 */
  unit: string;
  frequency: Frequency;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /**
     * @minimum 1
     * @maximum 3
     */
  priority: number;
}

export interface GoalUpdate {
  /** @minLength 1 */
  name?: string;
  category?: Category;
  /** @exclusiveMinimum 0 */
  targetValue?: number;
  /** @minLength 1 */
  unit?: string;
  frequency?: Frequency;
  startDate?: string;
  /** @nullable */
  endDate?: string | null;
  active?: boolean;
  /**
     * @minimum 1
     * @maximum 3
     */
  priority?: number;
}

export interface DailyGoal {
  id: number;
  goalId: number;
  name: string;
  category: Category;
  targetValue: number;
  currentValue: number;
  unit: string;
  status: GoalStatus;
  percent: number;
  priority: number;
}

export interface DailyGoalProgressInput {
  /** @minimum 0 */
  currentValue: number;
  status?: GoalStatus;
}

export interface CategoryContribution {
  category: Category;
  weight: number;
  completion: number;
  contribution: number;
}

export interface Dashboard {
  date: string;
  greeting: string;
  dailyCompletion: number;
  dailyScore: number;
  streak: number;
  focusMinutes: number;
  completedCount: number;
  totalCount: number;
  categoryContributions: CategoryContribution[];
  highlights: string[];
}

export interface Today {
  date: string;
  dailyCompletion: number;
  dailyScore: number;
  goals: DailyGoal[];
  categoryContributions: CategoryContribution[];
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Authentication required
 */
export type UnauthorizedResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;