# lib/api-zod/src/generated/types/categoryContribution.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LifeOS Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';

export interface CategoryContribution {
  category: Category;
  weight: number;
  completion: number;
  contribution: number;
}