# lib/api-zod/src/generated/types/dailyGoalProgressInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LifeOS Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { GoalStatus } from './goalStatus';

export interface DailyGoalProgressInput {
  /** @minimum 0 */
  currentValue: number;
  status?: GoalStatus;
}