# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LifeOS Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryContribution } from './categoryContribution';

export interface Dashboard {
  date: Date;
  greeting: string;
  dailyCompletion: number;
  dailyScore: number;
  streak: number;
  focusMinutes: number;
  completedCount: number;
  totalCount: number;
  categoryContributions: CategoryContribution[];
  highlights: string[];
}