# lib/api-zod/src/generated/types/frequency.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LifeOS Phase 1 API
 * OpenAPI spec version: 0.1.0
 */

export type Frequency = typeof Frequency[keyof typeof Frequency];


export const Frequency = {
  daily: 'daily',
  weekdays: 'weekdays',
  weekly: 'weekly',
} as const;