# lib/api-zod/src/generated/types/goal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LifeOS Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { Frequency } from './frequency';

export interface Goal {
  id: number;
  name: string;
  category: Category;
  targetValue: number;
  unit: string;
  frequency: Frequency;
  startDate: Date;
  /** @nullable */
  endDate: Date | null;
  active: boolean;
  /**
     * @minimum 1
     * @maximum 3
     */
  priority: number;
}