# lib/api-zod/src/generated/types/goalStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LifeOS Phase 1 API
 * OpenAPI spec version: 0.1.0
 */

export type GoalStatus = typeof GoalStatus[keyof typeof GoalStatus];


export const GoalStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  completed: 'completed',
  skipped: 'skipped',
} as const;