# lib/api-zod/src/generated/types/goalUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LifeOS Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { Frequency } from './frequency';

export interface GoalUpdate {
  /** @minLength 1 */
  name?: string;
  category?: Category;
  /** @exclusiveMinimum 0 */
  targetValue?: number;
  /** @minLength 1 */
  unit?: string;
  frequency?: Frequency;
  startDate?: Date;
  /** @nullable */
  endDate?: Date | null;
  active?: boolean;
  /**
     * @minimum 1
     * @maximum 3
     */
  priority?: number;
}