# lib/api-zod/src/generated/types/today.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LifeOS Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryContribution } from './categoryContribution';
import type { DailyGoal } from './dailyGoal';

export interface Today {
  date: Date;
  dailyCompletion: number;
  dailyScore: number;
  goals: DailyGoal[];
  categoryContributions: CategoryContribution[];
}